+++ apps/www/components/crumb-bar.tsx
@@ -3,7 +3,7 @@
 import Link from "next/link";
 import { usePathname } from "next/navigation";
 import * as React from "react";
-import { rasterComponents } from "@noord/raster";
+import { rasterComponents } from "@noordvc/raster";
 
 interface Crumb {
   label: string;